+++ checkpoint/orbax/checkpoint/experimental/emergency/test_utils.py
@@ -29,7 +29,6 @@ import jax
 import numpy as np
 import optax
 from orbax.checkpoint import args as args_lib
-from orbax.checkpoint import options as options_lib
 from orbax.checkpoint import test_utils
 from orbax.checkpoint import utils
 from orbax.checkpoint._src.handlers import pytree_checkpoint_handler
@@ -38,7 +37,6 @@ from orbax.checkpoint._src.multihost import multislice
 from orbax.checkpoint.experimental.emergency import checkpoint_manager
 from orbax.checkpoint.experimental.emergency import mesh_consistency
 from orbax.checkpoint.experimental.emergency import process_metadata_checkpoint_handler
-from orbax.checkpoint.path import step as step_lib
 
 PyTree = Any
 PyTreeCheckpointHandler = pytree_checkpoint_handler.PyTreeCheckpointHandler
@@ -149,11 +147,6 @@ class LocalCheckpointManagerTestBase:
       self.doubled_pytree = doubled_pytree
       self.mesh_tree = mesh_tree
       self.axes_tree = axes_tree
-
-      self.persistent_checkpoint_handler = PyTreeCheckpointHandler(
-          use_ocdbt=True,
-          use_zarr3=True,
-      )
 
       # make sure each process is working on different directories
       self.local_directory = epath.Path(
@@ -922,10 +915,6 @@ class CheckpointManagerTestBase:
           self.persistent_directory,
       )
 
-      self.persistent_checkpoint_handler = PyTreeCheckpointHandler(
-          use_ocdbt=True,
-          use_zarr3=True,
-      )
       test_utils.set_tensorstore_driver_for_test()
       test_utils.sync_global_processes('CheckpointManagerTest:setup_complete')
 
@@ -935,9 +924,8 @@ class CheckpointManagerTestBase:
           'CheckpointManagerTest:teardown_complete'
       )
 
-    def setup_pytree(self, replica_axis_index: int = 0):
+    def setup_pytree(self, global_mesh: jax.sharding.Mesh):
       """Create mesh and pytree."""
-      global_mesh = self.make_global_mesh(replica_axis_index)
       pytree = {
           'a': test_utils.create_sharded_array(
               np.arange(8), global_mesh, jax.sharding.PartitionSpec(None)
@@ -973,7 +961,7 @@ class CheckpointManagerTestBase:
         expectation: bool,
     ):
       """Test case."""
-      global_mesh, pytree = self.setup_pytree()
+      global_mesh, pytree = self.setup_pytree(self.make_global_mesh())
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       options = CheckpointManagerOptions(
           local=LocalCheckpointOptions(
@@ -1012,28 +1000,14 @@ class CheckpointManagerTestBase:
     )
     def test_global_max(self, inputs, expectation):
       """Test case."""
-      global_mesh, pytree = self.setup_pytree()
-      abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
-      options = CheckpointManagerOptions(
-          local=LocalCheckpointOptions(save_interval_steps=1, max_to_keep=1),
-          persistent=PersistentCheckpointOptions(
-              save_interval_steps=1, max_to_keep=4
-          ),
-      )
-      manager = CheckpointManager(
-          local_directory=self.local_directory,
-          persistent_directory=self.persistent_directory,
-          global_mesh=global_mesh,
-          abstract_state=abstract_state,
-          options=options,
-      )
-
       local_host_inputs = inputs[multihost.process_index()]
       if not isinstance(local_host_inputs, list):
         local_host_inputs = [local_host_inputs]
         expectation = [expectation]
       self.assertEqual(
-          manager._global_max(local_host_inputs),
+          checkpoint_manager._global_max(
+              local_host_inputs, checkpoint_manager._get_global_broadcast_fn()
+          ),
           expectation,
       )
 
@@ -1053,7 +1027,7 @@ class CheckpointManagerTestBase:
         expectation,
     ):
       """Test case."""
-      global_mesh, pytree = self.setup_pytree()
+      global_mesh, pytree = self.setup_pytree(self.make_global_mesh())
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       options = CheckpointManagerOptions(
           local=LocalCheckpointOptions(
@@ -1103,7 +1077,7 @@ class CheckpointManagerTestBase:
         total_steps: Optional[int] = 4,
     ):
       """Test case."""
-      global_mesh, pytree = self.setup_pytree()
+      global_mesh, pytree = self.setup_pytree(self.make_global_mesh())
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       options = CheckpointManagerOptions(
           local=LocalCheckpointOptions(
@@ -1144,7 +1118,7 @@ class CheckpointManagerTestBase:
     )
     def test_save(self, enable_async_checkpointing: bool):
       """Test case."""
-      global_mesh, pytree = self.setup_pytree()
+      global_mesh, pytree = self.setup_pytree(self.make_global_mesh())
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       options = CheckpointManagerOptions(
           local=LocalCheckpointOptions(save_interval_steps=3, max_to_keep=2),
@@ -1165,26 +1139,7 @@ class CheckpointManagerTestBase:
         manager.save(i, args=PyTreeSaveArgs(pytree))
       manager.wait_until_finished()
 
-      in_primary_slice = multislice.in_slice(
-          multihost.process_index(), global_mesh
-      )
-      if in_primary_slice:
-        persistent_multiprocessing_options = options_lib.MultiprocessingOptions(
-            primary_host=global_mesh.devices[0].flat[0].process_index,
-            active_processes=multihost.unique_processes_from_devices(
-                global_mesh.devices[0]
-            ),
-            barrier_sync_key_prefix='persistent',
-        )
-        persistent_manager = manager._make_persistent_checkpoint_manager(
-            persistent_multiprocessing_options
-        )
-        self.assertEmpty(step_lib.checkpoint_steps(self.local_directory))
-        self.assertEqual(persistent_manager.all_steps(), [5, 10, 15])
-      else:
-        local_manager = manager._make_local_checkpoint_manager()
-        self.assertNotEmpty(step_lib.checkpoint_steps(self.local_directory))
-        self.assertEqual(local_manager.all_steps(), [12, 15])
+      self.assertEqual(manager.all_steps(), [5, 10, 12, 15])
 
     @parameterized.parameters(
         (False, 0),
@@ -1195,7 +1150,9 @@ class CheckpointManagerTestBase:
         self, enable_async_checkpointing: bool, replica_axis_index: int
     ):
       """Test restore successfully."""
-      global_mesh, pytree = self.setup_pytree(replica_axis_index)
+      global_mesh, pytree = self.setup_pytree(
+          self.make_global_mesh(replica_axis_index)
+      )
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       pytree_double = test_utils.apply_function(pytree, lambda x: x * 2)
       options = CheckpointManagerOptions(
@@ -1208,7 +1165,7 @@ class CheckpointManagerTestBase:
       )
 
       with mock.patch.object(
-          CheckpointManager,
+          checkpoint_manager._MultisliceCheckpointManager,
           '_restore_from_persistent',
           autospec=True,
           return_value=None,
@@ -1237,7 +1194,9 @@ class CheckpointManagerTestBase:
     )
     def test_slice_setup(self, replica_axis_index: int):
       """Test restoration after slice swap (via global mesh)."""
-      global_mesh, pytree = self.setup_pytree(replica_axis_index)
+      global_mesh, pytree = self.setup_pytree(
+          self.make_global_mesh(replica_axis_index)
+      )
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       pytree_double = test_utils.apply_function(pytree, lambda x: x * 2)
       options = CheckpointManagerOptions(
@@ -1255,8 +1214,10 @@ class CheckpointManagerTestBase:
           options=options,
           abstract_state=abstract_state,
       ) as manager:
-        self.assertEqual(manager._persistent_primary_host, 0)
-        self.assertEqual(manager._local_primary_host, 2)
+        self.assertEqual(
+            manager._checkpoint_manager._persistent_primary_host, 0
+        )
+        self.assertEqual(manager._checkpoint_manager._local_primary_host, 2)
         slice_id = multislice.process_slice_id(
             multihost.process_index(),
             global_mesh,
@@ -1326,7 +1287,9 @@ class CheckpointManagerTestBase:
     )
     def test_slice_swap_check_steps(self, replica_axis_index: int):
       """Test step detection across slice swap (via mesh)."""
-      global_mesh, pytree = self.setup_pytree(replica_axis_index)
+      global_mesh, pytree = self.setup_pytree(
+          self.make_global_mesh(replica_axis_index)
+      )
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       options = CheckpointManagerOptions(
           local=LocalCheckpointOptions(save_interval_steps=1, max_to_keep=10),
@@ -1343,8 +1306,10 @@ class CheckpointManagerTestBase:
           options=options,
           abstract_state=abstract_state,
       ) as manager:
-        self.assertEqual(manager._persistent_primary_host, 0)
-        self.assertEqual(manager._local_primary_host, 2)
+        self.assertEqual(
+            manager._checkpoint_manager._persistent_primary_host, 0
+        )
+        self.assertEqual(manager._checkpoint_manager._local_primary_host, 2)
         slice_id = multislice.process_slice_id(
             multihost.process_index(),
             global_mesh,
@@ -1383,8 +1348,10 @@ class CheckpointManagerTestBase:
           options=options,
           abstract_state=abstract_state,
       ) as manager:
-        self.assertEqual(manager._persistent_primary_host, 2)
-        self.assertEqual(manager._local_primary_host, 0)
+        self.assertEqual(
+            manager._checkpoint_manager._persistent_primary_host, 2
+        )
+        self.assertEqual(manager._checkpoint_manager._local_primary_host, 0)
         slice_id = multislice.process_slice_id(
             multihost.process_index(),
             new_global_mesh,
@@ -1411,7 +1378,9 @@ class CheckpointManagerTestBase:
         self, enable_async_checkpointing: bool, replica_axis_index: int
     ):
       """Test restore successfully."""
-      global_mesh, pytree = self.setup_pytree(replica_axis_index)
+      global_mesh, pytree = self.setup_pytree(
+          self.make_global_mesh(replica_axis_index)
+      )
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       pytree_double = test_utils.apply_function(pytree, lambda x: x * 2)
       options = CheckpointManagerOptions(
@@ -1463,7 +1432,9 @@ class CheckpointManagerTestBase:
         self, enable_async_checkpointing: bool, replica_axis_index: int
     ):
       """Test case."""
-      global_mesh, pytree = self.setup_pytree(replica_axis_index)
+      global_mesh, pytree = self.setup_pytree(
+          self.make_global_mesh(replica_axis_index)
+      )
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       pytree_double = test_utils.apply_function(pytree, lambda x: x * 2)
       options = CheckpointManagerOptions(
@@ -1495,7 +1466,7 @@ class CheckpointManagerTestBase:
 
     def test_process_index_metadata(self):
       """Test case."""
-      global_mesh, pytree = self.setup_pytree()
+      global_mesh, pytree = self.setup_pytree(self.make_global_mesh())
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       with CheckpointManager(
           local_directory=self.local_directory,
@@ -1547,7 +1518,7 @@ class CheckpointManagerTestBase:
         return step in persistent_expectation
 
       max_to_keep = 10
-      global_mesh, pytree = self.setup_pytree()
+      global_mesh, pytree = self.setup_pytree(self.make_global_mesh())
       abstract_state = jax.tree.map(utils.to_shape_dtype_struct, pytree)
       options = CheckpointManagerOptions(
           local=LocalCheckpointOptions(
@@ -1593,6 +1564,8 @@ class CheckpointManagerTestBase:
         def __init__(self, devices, axis_names):
           self.devices = devices
           self.axis_names = axis_names
+          self.shape = mesh_shape
+          self.shape_tuple = tuple(mesh_shape)
 
       devices = np.arange(np.prod(mesh_shape)).reshape(mesh_shape)
       with mock.patch.object(jax.sharding, 'Mesh', wraps=FakeMesh):